mmaprototype: move rebalanceStores to its own file
Epic: CRDB-55052
