+++ pkg/kv/kvserver/allocator/mmaprototype/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
         "allocator.go",
         "allocator_state.go",
         "cluster_state.go",
+        "cluster_state_rebalance_stores.go",
         "constraint.go",
         "constraint_matcher.go",
         "doc.go",
